refactor(#998): co-locate ComplianceBanner with operator-fleet
The banner encodes an operator-fleet domain rule and imports
operator-fleet/api, but lived under operator-dashboard. Move it (and its
test) next to the rule it depends on. Pure move + import path update; no
logic change, banner still renders on the dashboard. Not lint-blocked
(src/vite is the deprecated-but-active tree).

Refs #998
